+++ modules/analysis-common/src/main/java/org/elasticsearch/analysis/common/RomanianAnalyzerProvider.java
@@ -17,29 +17,31 @@
  * under the License.
  */
 
-package org.elasticsearch.index.analysis;
+package org.elasticsearch.analysis.common;
 
 import org.apache.lucene.analysis.CharArraySet;
-import org.apache.lucene.analysis.tr.TurkishAnalyzer;
+import org.apache.lucene.analysis.ro.RomanianAnalyzer;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.index.IndexSettings;
+import org.elasticsearch.index.analysis.AbstractIndexAnalyzerProvider;
+import org.elasticsearch.index.analysis.Analysis;
 
-public class TurkishAnalyzerProvider extends AbstractIndexAnalyzerProvider<TurkishAnalyzer> {
+public class RomanianAnalyzerProvider extends AbstractIndexAnalyzerProvider<RomanianAnalyzer> {
 
-    private final TurkishAnalyzer analyzer;
+    private final RomanianAnalyzer analyzer;
 
-    public TurkishAnalyzerProvider(IndexSettings indexSettings, Environment env, String name, Settings settings) {
+    RomanianAnalyzerProvider(IndexSettings indexSettings, Environment env, String name, Settings settings) {
         super(indexSettings, name, settings);
-        analyzer = new TurkishAnalyzer(
-            Analysis.parseStopWords(env, settings, TurkishAnalyzer.getDefaultStopSet()),
+        analyzer = new RomanianAnalyzer(
+            Analysis.parseStopWords(env, settings, RomanianAnalyzer.getDefaultStopSet()),
             Analysis.parseStemExclusion(settings, CharArraySet.EMPTY_SET)
         );
         analyzer.setVersion(version);
     }
 
     @Override
-    public TurkishAnalyzer get() {
+    public RomanianAnalyzer get() {
         return this.analyzer;
     }
 }
