+++ server/src/main/java/org/elasticsearch/indices/analysis/PreBuiltAnalyzers.java
@@ -20,30 +20,12 @@ package org.elasticsearch.indices.analysis;
 
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.analysis.CharArraySet;
-import org.apache.lucene.analysis.ckb.SoraniAnalyzer;
 import org.apache.lucene.analysis.core.KeywordAnalyzer;
 import org.apache.lucene.analysis.core.SimpleAnalyzer;
 import org.apache.lucene.analysis.core.StopAnalyzer;
 import org.apache.lucene.analysis.core.WhitespaceAnalyzer;
-import org.apache.lucene.analysis.el.GreekAnalyzer;
-import org.apache.lucene.analysis.es.SpanishAnalyzer;
-import org.apache.lucene.analysis.fa.PersianAnalyzer;
-import org.apache.lucene.analysis.ga.IrishAnalyzer;
-import org.apache.lucene.analysis.hi.HindiAnalyzer;
-import org.apache.lucene.analysis.hu.HungarianAnalyzer;
-import org.apache.lucene.analysis.id.IndonesianAnalyzer;
-import org.apache.lucene.analysis.it.ItalianAnalyzer;
-import org.apache.lucene.analysis.lt.LithuanianAnalyzer;
-import org.apache.lucene.analysis.lv.LatvianAnalyzer;
-import org.apache.lucene.analysis.no.NorwegianAnalyzer;
-import org.apache.lucene.analysis.pt.PortugueseAnalyzer;
-import org.apache.lucene.analysis.ro.RomanianAnalyzer;
-import org.apache.lucene.analysis.ru.RussianAnalyzer;
 import org.apache.lucene.analysis.standard.ClassicAnalyzer;
 import org.apache.lucene.analysis.standard.StandardAnalyzer;
-import org.apache.lucene.analysis.sv.SwedishAnalyzer;
-import org.apache.lucene.analysis.th.ThaiAnalyzer;
-import org.apache.lucene.analysis.tr.TurkishAnalyzer;
 import org.elasticsearch.Version;
 import org.elasticsearch.indices.analysis.PreBuiltCacheFactory.CachingStrategy;
 
@@ -107,168 +89,6 @@ public enum PreBuiltAnalyzers {
         @Override
         protected Analyzer create(Version version) {
             Analyzer a = new ClassicAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    GREEK {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new GreekAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    HINDI {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new HindiAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    HUNGARIAN {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new HungarianAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    INDONESIAN {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new IndonesianAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    IRISH {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new IrishAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    ITALIAN {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new ItalianAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    LATVIAN {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new LatvianAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    LITHUANIAN {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new LithuanianAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    NORWEGIAN {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new NorwegianAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    PERSIAN {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new PersianAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    PORTUGUESE {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new PortugueseAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    ROMANIAN {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new RomanianAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    RUSSIAN {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new RussianAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    SORANI {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new SoraniAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    SPANISH {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new SpanishAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    SWEDISH {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new SwedishAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    TURKISH {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new TurkishAnalyzer();
-            a.setVersion(version.luceneVersion);
-            return a;
-        }
-    },
-
-    THAI {
-        @Override
-        protected Analyzer create(Version version) {
-            Analyzer a = new ThaiAnalyzer();
             a.setVersion(version.luceneVersion);
             return a;
         }
